# Improve performance for `Mysql2::Client#xquery`